+++ CHANGELOG.md
@@ -15,6 +15,7 @@ RC2 adds/fixes the following issue:
 ### Added
   - implemented STM32F1 analog input reading (#88)
   - implemented io_control analog reading function (#88)
+  - implemented SAMD21 analog reading function (#90)
 
 ### Changed
   - modified file structure and tinyUSB was modified and integrated to allow compiling code using Arduino IDE (#87)